backend/contract-test-scaffold/simulator/python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
sse-starlette==2.1.2
python-dateutil==2.9.0.post0
